+++ src/cross_check.py
@@ -36,7 +36,7 @@ def __extract_grid(grid: ConfiguredTextGrid, grid_name: str) -> dict[str, str]:
             (
                 i
                 for i in range(grid.text_grid.ncols)
-                if grid.text_grid.rows[grid.first_row][i] == grid.designator_col
+                if grid.text_grid.rows_raw()[grid.first_row][i] == grid.designator_col
             ),
             -1,
         )
@@ -45,7 +45,7 @@ def __extract_grid(grid: ConfiguredTextGrid, grid_name: str) -> dict[str, str]:
             (
                 i
                 for i in range(grid.text_grid.ncols)
-                if grid.text_grid.rows[grid.first_row][i] == grid.comment_col
+                if grid.text_grid.rows_raw()[grid.first_row][i] == grid.comment_col
             ),
             -1,
         )
@@ -69,8 +69,8 @@ def __extract_grid(grid: ConfiguredTextGrid, grid_name: str) -> dict[str, str]:
         raise ValueError(f"{grid_name} last row > number of rows")
 
     for row in range(first_row, last_row):
-        dsgn = grid.text_grid.rows[row][designator_col_idx]
-        cmnt = grid.text_grid.rows[row][comment_col_idx]
+        dsgn = grid.text_grid.rows_raw()[row][designator_col_idx]
+        cmnt = grid.text_grid.rows_raw()[row][comment_col_idx]
         dsgn = dsgn.split(',')
         # logging.debug(f"designators: '{dsgn}'")
         for d in dsgn: